% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_agol.R
\name{cleanData}
\alias{cleanData}
\title{A generalized function to start data cleaning}
\usage{
cleanData(
  raw_data,
  cols_to_remove = c("objectid", "InstanceName", "^app_.*", "GapsKey", "^Shrub.*"),
  id_replacement_names = c("globalid", "objectid", "parentglobalid")
)
}
\arguments{
\item{raw_data}{list of tabular data and metadata}

\item{cols_to_remove}{a vector containing the names of columns to remove}

\item{id_replacement_names}{a vector containing}
}
\value{
A list containing tabular data and metadata
}
\description{
A generalized function to start data cleaning
}
