% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_agol.R
\name{fetchItemInfo}
\alias{fetchItemInfo}
\title{Fetch information about any AGOL item}
\usage{
fetchItemInfo(item_id, root = "nps.maps.arcgis.com")
}
\arguments{
\item{item_id}{The unique item ID, found in the URL after "id="}

\item{root}{NPS users should keep the default. See \url{https://developers.arcgis.com/rest/users-groups-and-items/root.htm} for more information.}
}
\value{
A list containing the item information
}
\description{
Fetch information about any AGOL item
}
\examples{
\dontrun{
  info <- fetchItemInfo("373a421a19564b0996ff30d554be964b")
}

}
