% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_agol.R
\name{fetchPublicHostedFile}
\alias{fetchPublicHostedFile}
\title{Download a public file from AGOL}
\usage{
fetchPublicHostedFile(
  item_id,
  download_path = ".",
  root = "nps.maps.arcgis.com"
)
}
\arguments{
\item{item_id}{The unique item ID, found in the URL after "id="}

\item{download_path}{The folder path where the downloaded file should be saved}

\item{root}{NPS users should keep the default. See \url{https://developers.arcgis.com/rest/users-groups-and-items/root.htm} for more information.}
}
\value{
The file path of the downloaded file
}
\description{
Download a public file from AGOL
}
\examples{
\dontrun{
  fetchPublicHostedFile("31b357c26700467792cfb5f7778d72ea", "./my_files")
}

}
