% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_to_csv.R
\name{generateMetadataCSVs}
\alias{generateMetadataCSVs}
\title{Write data dictionaries to files}
\usage{
generateMetadataCSVs(
  metadata,
  dictionary_dir = here::here("data", "dictionary"),
  dictionary_filenames = c(tables = "data_dictionary_tables.txt", attributes =
    "data_dictionary_attributes.txt", categories = "data_dictionary_categories.txt"),
  verbose = FALSE
)
}
\arguments{
\item{metadata}{Metadata part of output of \code{fetchRawData()}}
}
\description{
Write data dictionaries to files
}
