% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_agol.R
\name{removeCols}
\alias{removeCols}
\title{A function to remove specified columns and associated metadata}
\usage{
removeCols(all_data, cols_to_remove = c("Editor", "Creator"))
}
\arguments{
\item{cols_to_remove}{a vector containing the names of columns to remove (default is creator and editor columns)}

\item{raw_data}{list of tabular data and metadata}
}
\description{
A function to remove specified columns and associated metadata
}
