% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_agol.R
\name{troubleshootMetadata}
\alias{troubleshootMetadata}
\title{A function to help troubleshoot differences between column names in data and metadata if querying data using metadata fails}
\usage{
troubleshootMetadata(all_data, returnAll = FALSE)
}
\arguments{
\item{all_data}{output of \code{fetchRawData()} containing data and metadata}

\item{returnAll}{Do you want to return all column names or only ones that don't have a match}
}
\value{
list of dfs containing data and metadata column names
}
\description{
A function to help troubleshoot differences between column names in data and metadata if querying data using metadata fails
}
