% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R, R/utils.R
\name{fetchRawData}
\alias{fetchRawData}
\title{Fetch tabular data and metadata from AGOL}
\usage{
fetchRawData(
  database_url,
  agol_username,
  agol_password = keyring::key_get(service = "AGOL", username = agol_username)
)

fetchRawData(
  database_url,
  agol_username,
  agol_password = keyring::key_get(service = "AGOL", username = agol_username)
)
}
\arguments{
\item{database_url}{Feature service URL}

\item{agol_username}{AGOL headless account username}

\item{agol_password}{AGOL headless account password (do not hard code this into your scripts!)}
}
\value{
A list containing tabular data and metadata

A list containing tabular data and metadata
}
\description{
Fetch tabular data and metadata from AGOL

Fetch tabular data and metadata from AGOL
}
