% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_to_csv.R
\name{generateMetadataCSVs}
\alias{generateMetadataCSVs}
\title{Write data dictionaries to files}
\usage{
generateMetadataCSVs(
  data,
  dictionary_dir = here::here("data", "dictionary"),
  dictionary_filenames = c(tables = "data_dictionary_tables.txt", attributes =
    "data_dictionary_attributes.txt", categories = "data_dictionary_categories.txt"),
  verbose = FALSE
)
}
\arguments{
\item{data}{output of \code{fetchRawData()}}

\item{dictionary_dir}{Folder to store data dictionaries in}

\item{dictionary_filenames}{Named list with names \code{c("tables", "attributes", "categories")} indicating what to name the tables, attributes, and categories data dictionaries. You are encouraged to keep the default names unless you have a good reason to change them.}

\item{verbose}{Output feedback to console?}
}
\description{
Write data dictionaries to files
}
