% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_agol.R
\name{removeCols}
\alias{removeCols}
\title{Remove specified columns and associated metadata}
\usage{
removeCols(
  all_data,
  cols_to_remove = c("CreationDate", "Creator", "EditDate", "Editor"),
  exact = TRUE
)
}
\arguments{
\item{all_data}{list of tabular data and metadata}

\item{cols_to_remove}{a vector containing the names of columns to remove (default is creator and editor columns)}

\item{exact}{Should the columns be matched exactly or should regular expressions be used?}
}
\description{
Remove specified columns and associated metadata
}
