% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R, R/utils.R
\name{setDataTypesFromMetadata}
\alias{setDataTypesFromMetadata}
\title{Set data types based on AGOL metadata}
\usage{
setDataTypesFromMetadata(raw_data)

setDataTypesFromMetadata(raw_data)
}
\arguments{
\item{raw_data}{list of tabular data and metadata}
}
\value{
A list containing tabular data and metadata

A list containing tabular data and metadata
}
\description{
Set data types based on AGOL metadata

Set data types based on AGOL metadata
}
